% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsen_POS_tags.R
\name{coarsen_POS_tags}
\alias{coarsen_POS_tags}
\title{Coarsen POS tags}
\usage{
coarsen_POS_tags(tag_vector)
}
\arguments{
\item{tag_vector}{A vector of POS tags.}
}
\value{
A vector of coarse tags.
}
\description{
Coarsens PTB or Petrov/Gimpel tags into one of five categories:
'A' = ADJ, 'D' = DET, 'P' = ADP, 'N' = NOUN, 'O' = all else
}

