% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsen_POS_tags.R
\name{coarsen_POS_tags}
\alias{coarsen_POS_tags}
\title{Coarsen POS tags}
\usage{
coarsen_POS_tags(tag_vector)
}
\arguments{
\item{tag_vector}{A vector of POS tags.}
}
\value{
A vector of coarse tags.
}
\description{
Coarsens PTB or Petrov/Gimpel coarse tags into one of five
categories:
'A' = adjective, 'D' = determiner, 'P' = preposition,
'N' = common/proper noun, 'O' = all else
}
\examples{
pos_tags <- c("VB", "JJ", "NN", "NN")
coarsen_POS_tags(pos_tags)
}

