% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ngram_filter.R
\name{extract_ngram_filter}
\alias{extract_ngram_filter}
\title{Extract phrase spans}
\usage{
extract_ngram_filter(pos_tags, regex, maximum_ngram_length,
  exclude_unigrams = FALSE)
}
\arguments{
\item{pos_tags}{A character vector of Penn TreeBank or Petrov/Gimpel
style tags.}

\item{regex}{The regular expression used to find phrases.}

\item{maximum_ngram_length}{The maximum length phrases returned. Defaults to
8. Increasing this number can greatly increase runtime.}

\item{exclude_unigrams}{Option to exclude all unigrams from spans returned.
Defaults to FALSE.}
}
\value{
A numeric matrix with two columns and rows equal to number of spans
matched. First column is span start, second is span end.
}
\description{
Takes a sequences of POS tags and a regex and returns spans
which match regex.
}

